Dev/vien2

+++ app/controllers/users_controller.py
@@ -68,11 +68,13 @@ def get_user(user_id):
 @jwt_required()
 def update_user(user_id):
     try:
+        current_user = User.objects.get(user_id=str(user_id))
+        email = request.json['email']
         old_password = request.json['old_password']
         new_password = generate_password_hash(request.json['new_password'])
-        check_password = User.objects.get(user_id=str(user_id)).password
+        check_password = current_user.password
         if check_password_hash(check_password, old_password):
-            User.objects(user_id=str(user_id)).update(password=new_password)
+            current_user.update(email=email, password=new_password)
             info = 'Update infomation for user_id:' + user_id
         else:
             info = "Wrong password!"
@@ -85,6 +87,11 @@ def update_user(user_id):
 @jwt_required()
 def delete_user(user_id):  
     try:
+        if User.objects(user_id=user_id).count() == 0:
+            return jsonify({"success": "False", "error": "user not does not exist"})
+        data_user = User.objects(user_id=user_id).first()
+        if data_user.public_token != None and data_user.public_token != 'delete':
+            Project.objects(public_token=data_user.public_token).delete()
         User.objects(user_id=user_id).delete()
         info = 'Delete user_id:' + user_id
     except Exception as e:
